use a fixed alias name for contains/items subqueries in SQL to avoid ambiguities
Currently `ArrayContainsFilter` uses the name of the property (properly
quoted) as the alias for the unnested array. Because that alias is not
of type `row`, it has to be referred as a plain alias instead of
qualified with a table. This creates an ambiguity if the alias is the
same as the name of some column, and postgres errors out.

This commit fixes this bug by making the alias fixed (`contents`).

Signed-off-by: Carol Schulze <carol@balena.io>