Update permissions filter code to use the word "actor" instead of "user"
This change updates the word "user" to "actor" to be more semantically
correct when working with permissions, since contracts that aren't users
can interact with the system.
This doesn't introduce any functional changes apart from making the
system more safe if the actor doesn't have a role field.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>