Remove fullTextSearch index on user slug
This index was added recently to facilitate searching for users by slug in a full text search query. But this isn't very sensible as it requires the user to include the `user-` prefix in the search term - which they will never (want to) do! So I'm removing the fullTextSearch index and the CreateView lens will manually search for the user by slug using a regex search.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>