Ensure that backend types are included in compiled build
With our tsc configuration, `*.d.ts` files are ignored by the compiler
and won't be copied across to the output directory, this causes problems
when downstream projects are trying to use types defined here.
This change simply changes the extension to `.ts`, causing the file to
be included in the build output.

See https://stackoverflow.com/a/56440335 for more detail

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>