Automatically generate slugs when inserting contracts
Previously slugs were required, and `action-create-card` would
automatically generate the slug. This change moves the slug generation
into the core (which I think makes more sense) and also improves the
slug generation by removing stop words and only using a short uuid if
there is a name available.
Note that a test was removed here, as the assertion it made is now
covered by the type signature on the `insertCard` method.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>