Drop the slugify module dependency in favor of simple inline code
Using the @sindresorhus/slugify module was causing a lot of headaches as
it is shipped as an esm module and requires additional compiler and jest
config. The additional effort wasn't worth the convenience, so I decided
to replace it with a simple inline statement that is functionally the
same.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>