Safely handle simultaneous creation of identical indexes
If two instances run this function simultaneously, it's possible to enter a race condition
where the index creation will fail with an error, as one instance will create the index
before the other. Ideally we want a unified way of handling system bootstrapping, but for now we
can simply retry once on a failure and check again to see if the index already exists.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>