Rationalize the use of a "context"
Currently, many functions take an argument called `context`. This is
simply an arbitrary object that should be passed as first argument to
calls into the logging or assert libraries. This commit introduces a
`Context` class that encapsulates the logging context. It also exposes
logging and assert functions so that, for example, calls like
`logger.info(context, ...` are replaced by `context.info(...`.

In the future the `Context` class will carry more than just the log
context.

Change-type: patch
Signed-off-by: Carol Schulze <carol@balena.io>