Export a `testUtils` module with utilities for testing against the core
`testUtils` has a `TestContext` that contains the three pieces necessary
to run tests against the core:

- The log context.
- The admin session token.
- The `Kernel` object.

`generateRandomId` (notice the change in capitalization) and
`generateRandomSlug` are not part of the context anymore. They are now
functions exported from the `testUtils` module.

Other test contexts that depend on the core's should extend
`TestContext` and use `newContext` and `destroyContext`.

Change-type: minor
Signed-off-by: Carol Schulze <carol@balena.io>