mixins: Add unit test for .withEvents()
This change adds a simple unit test to sanity check the formula used for
aggregating values in the `withEvents()` mixin. This change highlights
that the core should probably not be providing mixins, especialy with
functionality that it doesn't provide or handle directly ($$formula
fields). In the future we should consider hoisting mixins into the
worker module to improve feature seperation and make the system less
brittle.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>