Remove session contract indirection when finding identity of actor
With this change, autumndb sessions are now expected to be objects
containing the actor and an optional scope directly instead of a string
referring to the session contract. Validating session contracts
should now be handled by the implementing app layer.
This means that autumndb is only dealing with identity, instead of
partially handling auth.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>