Do not create `authentication-password` contracts for users that don't have what looks like a valid hash
Setting the hash to an invalid hash disables login indirectly. This is
usually done by setting the hash to the constant `PASSWORDLESS` but the
same effect can be achieved with other strings. Some tests and a few
contracts on live do this so we need to check if the hash at least looks
valid.

Change-type: patch
Signed-off-by: Carol Schulze <carol@balena.io>