Support streaming for linked contracts
This change multiplexes streams, creating additional streams that invert
the query to stream for linked data.
The query is introspected and if it traverse links, the corresponding
relationship is found and used to generate and inversion of the query,
which is then streamed.
When the inverse stream gets a result, the even is inverted again to
produce the expected event emission.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>